added tests to makefile and reorganized the project.
