+++ .gitignore
@@ -1,4 +1,3 @@
-cloudsrv
-cloudcli
+*.elf
 *.o
