+++ include/resource.h
