+++ src/client.c
@@ -5,7 +5,7 @@
 #include "network.h"
 #include "message.h"
 
-void print_help()
+void client_print_help()
 {
 	printf("USAGE: cloudcli <hostname> <command>\n");
 }
@@ -14,7 +14,7 @@ void print_help()
  *	This function initialize the client's socket
  *	and connects to the server.
  **/
-int init_client(char *hostname, short port_addr)
+int client_init(char *hostname, short port_addr)
 {
 	int 				sockfd;
   	struct sockaddr_in 	srv_addr;
@@ -50,7 +50,7 @@ out_err:
 	return -1;
 }
 
-int do_client(int sockfd, char* text) 
+int client_do(int sockfd, char* text) 
 {
 	message msg;
 	ssize_t ret = 0;
@@ -65,37 +65,6 @@ int do_client(int sockfd, char* text)
 		return -1;
 	}
 
-	return 0;
-}
-
-int main(int argc, char *argv[])
-{
-  	int 					sockfd 		= 0;
-  	short 					port_addr 	= 0;
-  	char					*hostname	= 0;
-  	char 					text[MSG_TEXT_SIZE] = {0};
-
-	if (argc < 3) {
-		print_help();
-		goto out_ret;
-	}
-
-	hostname = argv[1];
-
-	snprintf(text, sizeof(text), "%s", argv[2]);
-	for(int i = 3; i < argc; i++) {
-		snprintf(text, sizeof(text), "%s %s", text, argv[i]);
-	}
-
-	sockfd = init_client(hostname, SERVER_PORT);
-
-	printf("CONNECTED\n");
-
-	do_client(sockfd, text);
-
-out:
-	close(sockfd);
-out_ret:
 	return 0;
 }
