+++ tests/queue_test.cc
@@ -1,0 +1,6 @@
+#include <gtest/gtest.h>
+
+TEST(test_group, test1)
+{
+	ASSERT_EQ(1,1);
+}
